Colocando os links da imagens

+++ src/database/migrations/20210429202710_table_praias.js
@@ -7,7 +7,7 @@ exports.up = function(knex) {
     table.string('foto').notNullable();
     table.string('lat').notNullable();
     table.string('long').notNullable();
-    table.decimal('avaliacao').notNullable();
+    table.float('avaliacao').notNullable();
 
     table.string('municipios').notNullable();
     table.foreign('municipios').references('municipios.id_municipios');
